Fix generation of the forward header for QFunctionPointer
The forward header for QFunctionPointer was being generated by accident,
apparently syncqt is confused because of some of the preceding lines. If
the lines are rearranged or removed, the header won't be generated. Not
sure if there's much value in generating a header for it, but at least
one of Qt sources (qeglfsemulatorintegration.h) is using it in an
include statment, and there's a chance it's used in the user code as
well.

Change-Id: Iced505881ddb07e7a8336fcdfb7d3e5397c6466c
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
(cherry picked from commit 8b2145463b38291dfa0ca2acf4a85552b7383a60)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>