QDBusMetaObjectGenerator: replace a few more int
The meta object is actually uint, even though we end up treating it as
int elsewhere. I've also replaced one flags use with quint32.

That leaves int in this file for meta type IDs only. Replacing them with
QMetaTypeInterface pointers is left as an exercise for the reader.

Task-number: QTBUG-103548
Change-Id: Ie4bb662dcb274440ab8bfffd17096e0f85e4ea5a
Reviewed-by: Marc Mutz <marc.mutz@qt.io>