QUnicodeTools: fix data race in initialization of libthai symbols
The facilities of qunicodetools.cpp are not limited to the GUI thread,
so initialization must be thread-safe.

The old code wasn't, though, and contained several data races

- non-atomic initialized was read while another thead may write it
- th_brk and th_next_cell were read while another thead may write them

Fix by using Double-Checked Locking. This also prepares the code for
an eventual port to th_brk_find_breaks() (th_brk is deprecated).

The function pointers don't need to be atomic, because all reads from
them are guaranteed to happen-after the writes to them (as long as all
users call init_libthai() and don't proceeed if it returns false; this
could be ensured by returning a struct with the function pointers from
init_libthai() instead of maintaining them as statically-visible
globals, but that's outsize the scope of this patch).

As a drive-by, remove a pointless static_cast<int>(~~int expression~~).

Manual conflict resolutions:
  - fixed line conflcits d/t _L1 vs. QLatin1String
  - dropped Q_CONSTINIT macro (only available in 6.4+)

Fixes: QTBUG-105543
Change-Id: I492acd7e9a257e5c4b91f576e9bc448b6bb96ad1
Reviewed-by: Ievgenii Meshcheriakov <ievgenii.meshcheriakov@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@gmail.com>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
(cherry picked from commit 7d021e3173a03d384ac4b987545f8fdf52593a61)
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>