Repopulate theme fonts after invalidating CoreText font database
The cross platform font database maintains the font descriptors for
the system (theme) fonts, so when the font database is invalidated
we need to clear our list of theme fonts so that the next call to
QCoreTextFontDatabase::populateThemeFonts() will re-populate the
system font descriptors.

This issue was fixed in dev as a side effect of the change in
d84018616a24fa66502cf62f5ebbcf5ed4ca0742.

Change-Id: Iae0b1a62cca6326f7edf16d22daa94b86eaa0f2c
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
(cherry picked from commit 1ef1dca5458bed65a19e94ab62b50917ed46645d)
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>