Work round macOS's omission of en_DE from its own uiLanguages()
When the system locale is en_DE, macOS seems to think we should use
en_GB as the right translation. While that probably is a sensible
choice in the absence of an en_DE translation, we should definitely
use the en_DE translation if available, especially if en_GB isn't
available (which lead to a fall-back to de_DE, given later entries in
macOS's list). So prepend the system locale's own pcp47Name() if it
(isn't the C locale and) is missing from what we would otherwise have
used for uiLanguages(), after likely sub-tag perturbations.

Add a test simulating (some approximation to) what macOS was doing
that would have caught this case; and add a scope-guard reporter to
the test to report what shows up when lists don't match.

Conflicts resolved: 6.4 has a fix for a related issue, that means the
last two languages in the new test-case are swapped. Qt 6 has richer
infrastructure in QLocaleId that's missing in 5.15, so I had to port
the uses of it to the facilities present in 5.15.

Fixes: QTBUG-104930
Change-Id: I116234708067e1717d9157aebc84da76e04a9f38
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 753bfdf6a1336cc296e5fc8175aedd000c2cc013)
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>